detection_app/requirements.txt
Flask
numpy
pillow
requests
pyopenssl
